Add passability to stock short 1.875m to 1.25m adapter
Given that the stock rockomax adapters are passable, and given that the 1.875 to 1.25m is literally just a hollow conic section, I can't see any reason not to assume it can be pressurized and let people travel through it.  That's why I'm suggesting it go here rather than the Freedom cfg.  Otherwise, transits between 1.875 and 1.25 modules are pretty tough, as all of the longer adapters look like they have fuel tanks or something in there.