Bug #112 - potential fix for handling surface-attached parts.
When a part is surface-attached and later the vessel is re-rooted so
that the surface-attached part is now "upstream" of the part it is
surface-attached to, CLS cannot correctly work out the Living Spaces.

This is because there is no valid attach-node from the surface-attached
part to the part it is surface-attached to.

This patch also looks at the surface-attached -connection from the other
direction to determine whether the connection should be passable or
not.